combining apps into a single app
