--- apps/emss-2/Dashboard/dashboard.html
@@ -1,37 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<div class="col-xl-12 col-lg-12 col-md-12 col-sm-12 col-12">
-    <div class="row gutters">
-        <div class="col-xl-6 col-lg-6 col-md-6 col-sm-6 col-12">
-            <div class="card">
-                <div class="card-header">
-                    <div class="card-title">Manning Summary</div>
-                </div>
-                <div class="card-body">
-                    <form method="post" action="{% url 'manning_summary' %}">
-                        {% csrf_token %}
-                        <div class="form-group">
-                            <label for="projects">Projects</label>
-                            <select name="projects" id="projects" multiple>
-                                {% for project in projects %}
-                                    <option value="{{ project.ProjectId }}">{{ project.ProjectName }}</option>
-                                {% endfor %}
-                            </select>
-                        </div>
-                        <div class="form-group">
-                            <label for="jobs">Job Title</label>
-                            <select name="jobs" id="jobs" multiple>
-                                {% for job in jobs %}
-                                    <option value="{{ job.JobId }}">{{ job.JobName }}</option>
-                                {% endfor %}
-                            </select>
-                        </div>
-                        <button type="submit" class="btn btn-primary">Filter</button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
