--- bulkupload/templates/bulk_upload_employee.html
@@ -1,10 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<h2>Bulk Upload Employee</h2>
-<form method="post" enctype="multipart/form-data">
-    {% csrf_token %}
-    {{ form.as_p }}
-    <button type="submit">Upload</button>
-</form>
-{% endblock %}
