+++ dashboard/apps.py
@@ -4,3 +4,12 @@ from django.apps import AppConfig
 class DashboardConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'dashboard'
+
+#regstraton 
+
+from django.apps import AppConfig
+
+
+class RegistrationConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'registration'
