+++ dashboard/models.py
@@ -36,3 +36,15 @@ class WorkOrder(models.Model):
     ClosedInMoreThanFiveDay = models.IntegerField()
     TotalWorkOrders = models.IntegerField()
     CompletionRate = models.FloatField()
+
+
+#Registration
+from django.db import models
+
+class UserCredentials(models.Model):
+    name = models.CharField(max_length=100)
+    username = models.EmailField(unique=True)
+    password = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.name
