+++ dashboard/urls.py
@@ -5,3 +5,15 @@ from . import views
 urlpatterns = [
     path('', views.dashboard, name='dashboard'),
 ]
+
+#Regstarton
+
+# EMS/urls.py
+from django.contrib import admin
+from django.urls import path
+from registration import views
+
+urlpatterns = [
+    path('admin/', admin.site.urls),
+    path('register/', views.registration_view, name='register'),
+]
