--- registration/models.py
@@ -1,9 +1,0 @@
-from django.db import models
-
-class UserCredentials(models.Model):
-    name = models.CharField(max_length=100)
-    username = models.EmailField(unique=True)
-    password = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.name
