--- registration/templates/regidtration.html
@@ -1,70 +1,0 @@
-<!-- registration/templates/registration/registration.html -->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Registration</title>
-    <!-- External CSS libraries -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Open+Sans:400,300,600,700,800%7CPoppins:400,500,700,800,900%7CRoboto:100,300,400,400i,500,700">
-    <!-- Custom Stylesheet -->
-    <link rel="stylesheet" href="{% static 'registration/css/style.css' %}">
-</head>
-<body>
-    <div class="login-20">
-        <div class="container-fluid">
-            <div class="row">
-                <div class="col-xl-4 col-lg-5 col-md-12 bg-color-10">
-                    <div class="form-section">
-                        <div class="logo clearfix">
-                            <a href="#">
-                                <img src="{% static 'registration/images/EMSLogo.jpg' %}" alt="logo">
-                            </a>
-                        </div>
-                        <h3>Sign up your account</h3>
-                        <div class="login-inner-form">
-                            <form method="post">
-                                {% csrf_token %}
-                                <div class="form-group form-box">
-                                    {{ form.name }}
-                                    {% if form.name.errors %}
-                                        <div class="text-danger">{{ form.name.errors }}</div>
-                                    {% endif %}
-                                </div>
-                                <div class="form-group form-box">
-                                    {{ form.username }}
-                                    {% if form.username.errors %}
-                                        <div class="text-danger">{{ form.username.errors }}</div>
-                                    {% endif %}
-                                </div>
-                                <div class="form-group form-box">
-                                    {{ form.password }}
-                                    {% if form.password.errors %}
-                                        <div class="text-danger">{{ form.password.errors }}</div>
-                                    {% endif %}
-                                </div>
-                                <div class="form-group form-box">
-                                    {{ form.confirm_password }}
-                                    {% if form.confirm_password.errors %}
-                                        <div class="text-danger">{{ form.confirm_password.errors }}</div>
-                                    {% endif %}
-                                </div>
-                                <div class="form-group">
-                                    <button type="submit" class="btn-md btn-theme btn-block">Submit</button>
-                                </div>
-                            </form>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-xl-8 col-lg-7 col-md-12 bg-img none-992">
-                    <div class="info" style="margin-top:-30px;margin-bottom:-20px;">
-                        <img src="{% static 'registration/images/LoginRightBanner.jpg' %}" alt="" width="900">
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
